Tree_Project/requirements.txt
# Python dependencies for DRAC Graph Classifier

# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0

# Graph neural networks
torch-geometric>=2.3.0
torch-scatter>=2.1.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Image processing
Pillow>=9.0.0
scikit-image>=0.19.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
